feat: add product measurement ranking
- Adds a `rank` field to `WeightMeasurementEntity` to rank measurements within a product.
- Updates `ProductWeightMeasurementJunction` to include measurement rank.
- Implements ranking for product measurements in `AddFoodRepositoryImpl`.
- Updates `observeProductIdsWithMeasurementIds` in `AddFoodDao` to order by rank.
- Adds a `DEFAULT_RANK` for when rank is null.
- Adds a new `MeasurementWithRank` data class to hold the measurement id with rank.
- Modifies `HolderKey` to use `Float` for `extraId` instead of `Int`.
- Updates `SearchHome` to use rank in list keys and pass rank to `ProductSearchListItem`.
- Updates database schema to version 2 with changes to `WeightMeasurementEntity`.

+++ core/src/main/java/com/maksimowiczm/foodyou/core/feature/addfood/ui/search/SearchViewModel.kt
@@ -200,5 +200,5 @@ class SearchViewModel(
 
 data class HolderKey(
     val productId: Long,
-    val extraId: Int
+    val extraId: Float
 )
